# openapi.committed.json
{
  "components": {
    "schemas": {
...
      "EstimateTaxRequest": {
        "properties": {
          "account_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Account Id"
          },
          "filing_status": {
            "default": "single",
            "enum": [
              "single",
              "married_joint",
              "married_separate",
              "head_of_household"
            ],
            "title": "Filing Status",
            "type": "string"
          },
          "include_unrealized": {
            "default": false,
            "title": "Include Unrealized",
            "type": "boolean"
          },
          "tax_year": {
            "title": "Tax Year",
            "type": "integer"
          }
        },
        "required": [
          "tax_year"
        ],
        "title": "EstimateTaxRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RecordPaymentRequest": {
        "properties": {
          "confirm": {
            "title": "Confirm",
            "type": "boolean"
          },
          "payment_amount": {
            "exclusiveMinimum": 0.0,
            "title": "Payment Amount",
            "type": "number"
          },
          "quarter": {
            "enum": [
              "Q1",
              "Q2",
              "Q3",
              "Q4"
            ],
            "title": "Quarter",
            "type": "string"
          },
          "tax_year": {
            "title": "Tax Year",
            "type": "integer"
          }
        },
        "required": [
          "quarter",
          "tax_year",
          "payment_amount",
          "confirm"
        ],
        "title": "RecordPaymentRequest",
        "type": "object"
      },
...
      "SimulateTaxRequest": {
        "properties": {
          "account_id": {
            "title": "Account Id",
            "type": "string"
          },
          "action": {
            "enum": [
              "sell",
              "cover"
            ],
            "title": "Action",
            "type": "string"
          },
          "cost_basis_method": {
            "default": "fifo",
            "enum": [
              "fifo",
              "lifo",
              "specific_id",
              "avg_cost"
            ],
            "title": "Cost Basis Method",
            "type": "string"
          },
          "price": {
            "exclusiveMinimum": 0.0,
            "title": "Price",
            "type": "number"
          },
          "quantity": {
            "minimum": 1.0,
            "title": "Quantity",
            "type": "integer"
          },
          "ticker": {
            "title": "Ticker",
            "type": "string"
          }
        },
        "required": [
          "ticker",
          "action",
          "quantity",
          "price",
          "account_id"
        ],
        "title": "SimulateTaxRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "WashSaleRequest": {
        "properties": {
          "account_id": {
            "title": "Account Id",
            "type": "string"
          },
          "date_range_end": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Range End"
          },
          "date_range_start": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Range Start"
          },
          "ticker": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ticker"
          }
        },
        "required": [
          "account_id"
        ],
        "title": "WashSaleRequest",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Trading journal REST API",
    "title": "Zorivest API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/tax/audit": {
      "post": {
        "description": "Run transaction audit checks.",
        "operationId": "run_tax_audit_api_v1_tax_audit_post",
        "parameters": [
          {
            "in": "query",
            "name": "account_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Account Id"
            }
          },
          {
            "in": "query",
            "name": "tax_year",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tax Year"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Run Tax Audit Api V1 Tax Audit Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Tax Audit",
        "tags": [
          "tax"
        ]
      }
    },
    "/api/v1/tax/estimate": {
      "post": {
        "description": "Compute overall tax estimate.",
        "operationId": "estimate_tax_api_v1_tax_estimate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EstimateTaxRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Estimate Tax Api V1 Tax Estimate Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Estimate Tax",
        "tags": [
          "tax"
        ]
      }
    },
    "/api/v1/tax/harvest": {
      "get": {
        "description": "Scan portfolio for harvestable losses.",
        "operationId": "harvest_losses_api_v1_tax_harvest_get",
        "parameters": [
          {
            "in": "query",
            "name": "account_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Account Id"
            }
          },
          {
            "in": "query",
            "name": "min_loss_threshold",
            "required": false,
            "schema": {
              "default": 0.0,
              "title": "Min Loss Threshold",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "exclude_wash_risk",
            "required": false,
            "schema": {
              "default": false,
              "title": "Exclude Wash Risk",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Harvest Losses Api V1 Tax Harvest Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Harvest Losses",
        "tags": [
          "tax"
        ]
      }
    },
    "/api/v1/tax/lots": {
      "get": {
        "description": "List tax lots with basis, holding period, and gain/loss.",
        "operationId": "get_tax_lots_api_v1_tax_lots_get",
        "parameters": [
          {
            "in": "query",
            "name": "account_id",
            "required": true,
            "schema": {
              "title": "Account Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "ticker",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ticker"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "default": "all",
              "enum": [
                "open",
                "closed",
                "all"
              ],
              "title": "Status",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sort_by",
            "required": false,
            "schema": {
              "default": "acquired_date",
              "enum": [
                "acquired_date",
                "cost_basis",
                "gain_loss"
              ],
              "title": "Sort By",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Tax Lots Api V1 Tax Lots Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Tax Lots",
        "tags": [
          "tax"
        ]
      }
    },
    "/api/v1/tax/lots/{lot_id}/close": {
      "post": {
        "description": "Close a specific tax lot (mark as sold).",
        "operationId": "close_tax_lot_api_v1_tax_lots__lot_id__close_post",
        "parameters": [
          {
            "in": "path",
            "name": "lot_id",
            "required": true,
            "schema": {
              "title": "Lot Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Close Tax Lot Api V1 Tax Lots  Lot Id  Close Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Close Tax Lot",
        "tags": [
          "tax"
        ]
      }
    },
    "/api/v1/tax/lots/{lot_id}/reassign": {
      "put": {
        "description": "Reassign cost basis method for a lot.",
        "operationId": "reassign_lot_basis_api_v1_tax_lots__lot_id__reassign_put",
        "parameters": [
          {
            "in": "path",
            "name": "lot_id",
            "required": true,
            "schema": {
              "title": "Lot Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Body",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Reassign Lot Basis Api V1 Tax Lots  Lot Id  Reassign Put",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reassign Lot Basis",
        "tags": [
          "tax"
        ]
      }
    },
    "/api/v1/tax/quarterly": {
      "get": {
        "description": "Compute quarterly estimated tax payment obligations.",
        "operationId": "get_quarterly_estimate_api_v1_tax_quarterly_get",
        "parameters": [
          {
            "in": "query",
            "name": "quarter",
            "required": true,
            "schema": {
              "enum": [
                "Q1",
                "Q2",
                "Q3",
                "Q4"
              ],
              "title": "Quarter",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "tax_year",
            "required": true,
            "schema": {
              "title": "Tax Year",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "estimation_method",
            "required": false,
            "schema": {
              "default": "annualized",
              "enum": [
                "annualized",
                "actual",
                "prior_year"
              ],
              "title": "Estimation Method",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Quarterly Estimate Api V1 Tax Quarterly Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Quarterly Estimate",
        "tags": [
          "tax"
        ]
      }
    },
    "/api/v1/tax/quarterly/payment": {
      "post": {
        "description": "Record an actual quarterly tax payment. Requires confirm=true.",
        "operationId": "record_quarterly_tax_payment_api_v1_tax_quarterly_payment_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordPaymentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Record Quarterly Tax Payment Api V1 Tax Quarterly Payment Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Record Quarterly Tax Payment",
        "tags": [
          "tax"
        ]
      }
    },
    "/api/v1/tax/simulate": {
      "post": {
        "description": "Pre-trade what-if tax simulation.",
        "operationId": "simulate_tax_impact_api_v1_tax_simulate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimulateTaxRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Simulate Tax Impact Api V1 Tax Simulate Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Simulate Tax Impact",
        "tags": [
          "tax"
        ]
      }
    },
    "/api/v1/tax/wash-sales": {
      "post": {
        "description": "Detect wash sale chains/conflicts within 30-day windows.",
        "operationId": "find_wash_sales_api_v1_tax_wash_sales_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WashSaleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Find Wash Sales Api V1 Tax Wash Sales Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Find Wash Sales",
        "tags": [
          "tax"
        ]
      }
    },
    "/api/v1/tax/wash-sales/scan": {
      "post": {
        "description": "Trigger a full wash sale scan across all accounts.",
        "operationId": "scan_wash_sales_api_v1_tax_wash_sales_scan_post",
        "parameters": [
          {
            "in": "query",
            "name": "account_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Account Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Scan Wash Sales Api V1 Tax Wash Sales Scan Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Scan Wash Sales",
        "tags": [
          "tax"
        ]
      }
    },
    "/api/v1/tax/ytd-summary": {
      "get": {
        "description": "Year-to-date tax summary dashboard data.",
        "operationId": "get_ytd_tax_summary_api_v1_tax_ytd_summary_get",
        "parameters": [
          {
            "in": "query",
            "name": "tax_year",
            "required": true,
            "schema": {
              "title": "Tax Year",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "account_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Account Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Ytd Tax Summary Api V1 Tax Ytd Summary Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Ytd Tax Summary",
        "tags": [
          "tax"
        ]
      }
    },
...
  },
...
}